# specs/001-water-plant-discovery/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowGrid Water Plant Discovery API
  description: |
    RESTful API for FlowGrid Water Plant Discovery Platform V1.

    ## Authentication
    - Consumer App: JWT Bearer token (username/password login)
    - Owner App: JWT Bearer token (phone/password login)
    - Admin Dashboard: JWT Bearer token (username/password login)

    ## Rate Limiting
    - Public endpoints: 100 requests/minute per IP
    - Authenticated endpoints: 500 requests/minute per user

    ## Response Format
    All responses follow the standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "timestamp": "...", "requestId": "..." }
    }
    ```

    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": { ... }
      }
    }
    ```
  version: 1.0.0
  contact:
    name: FlowGrid API Support
    email: api@flowgrid.io
servers:
  - url: https://api.flowgrid.io/v1
    description: Production server
  - url: https://staging-api.flowgrid.io/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /consumer/auth/login:
    post:
      tags:
        - Consumer Auth
      summary: Consumer login
      description: Authenticate consumer with username and password
      operationId: consumerLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /consumer/auth/refresh:
    post:
      tags:
        - Consumer Auth
      summary: Refresh access token
      operationId: consumerRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # PLANTS (PUBLIC)
  # ============================================================================
...
  /owner/auth/register:
    post:
      tags:
        - Owner Auth
      summary: Register new owner
      description: Register a new plant owner account
      operationId: ownerRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerAuthResponse'
        '409':
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /owner/auth/login:
    post:
      tags:
        - Owner Auth
      summary: Owner login
      description: Authenticate owner with phone and password
      operationId: ownerLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /owner/auth/refresh:
    post:
      tags:
        - Owner Auth
      summary: Refresh owner access token
      operationId: ownerRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /owner/auth/password/reset:
    post:
      tags:
        - Owner Auth
      summary: Request password reset
      description: Send password reset OTP to owner's phone
      operationId: ownerRequestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  pattern: '^\+91[6-9]\d{9}$'
                  example: '+919876543210'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /owner/auth/password/confirm:
    post:
      tags:
        - Owner Auth
      summary: Confirm password reset
      description: Verify OTP and set new password
      operationId: ownerConfirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - otp
                - newPassword
              properties:
                phone:
                  type: string
                  pattern: '^\+91[6-9]\d{9}$'
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # OWNER PROFILE
  # ============================================================================
...
  /owner/notifications:
    get:
      tags:
        - Owner Profile
      summary: Get notifications
      description: Get owner's notifications (verification status changes, etc.)
      operationId: getOwnerNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unread
          in: query
          description: Filter to unread only
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /owner/notifications/{notificationId}/read:
    post:
      tags:
        - Owner Profile
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # CONSUMER CHAT
  # ============================================================================
...
  /admin/auth/login:
    post:
      tags:
        - Admin Auth
      summary: Admin login
      description: Authenticate admin with username and password
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/auth/refresh:
    post:
      tags:
        - Admin Auth
      summary: Refresh admin access token
      operationId: adminRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # ADMIN DASHBOARD
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Invalid or expired token'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'NOT_FOUND'
              message: 'Resource not found'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ========== Auth ==========
    ConsumerLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 1
    OwnerLoginRequest:
      type: object
      required:
        - phone
        - password
      properties:
        phone:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
          example: '+919876543210'
        password:
          type: string
          minLength: 1
    OwnerRegisterRequest:
      type: object
      required:
        - name
        - phone
        - password
        - plantName
        - address
        - latitude
        - longitude
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: 'Ramesh Kumar'
        phone:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
          example: '+919876543210'
        password:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        plantName:
          type: string
          minLength: 2
          maxLength: 200
          example: 'Pure Water Plant'
        address:
          type: string
          example: '123 Main Road, Hyderabad'
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        operatingHours:
          type: string
          maxLength: 200
          example: '8 AM - 8 PM, Mon-Sat'
    AdminLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
              description: Access token expiry in seconds
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
    OwnerAuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
            owner:
              $ref: '#/components/schemas/OwnerBasic'
            plant:
              $ref: '#/components/schemas/PlantBasic'
    AdminAuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
            admin:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                role:
                  type: string
                  enum: [super_admin, admin]

    # ========== Owner ==========
    OwnerBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
...
    PlantBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isOpen:
          type: boolean
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [verification_approved, verification_rejected, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            notifications:
              type: array
              items:
                $ref: '#/components/schemas/Notification'
            unreadCount:
              type: integer
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'

    # ========== Chat/Messaging ==========
...